move: hub 관련 dto 디렉토리 정리
