+++ hub/src/main/java/com/klp/hub/hub/presentation/dto/request/hub/UpdateHubRequest.java
@@ -1,4 +1,4 @@
-package com.klp.hub.hub.presentation.dto.request;
+package com.klp.hub.hub.presentation.dto.request.hub;
 
 import com.klp.hub.hub.application.command.UpdateHubCommand;
