+++ hub/src/main/java/com/klp/hub/hub/presentation/dto/response/hub/GetHubDetailResponse.java
@@ -1,8 +1,8 @@
-package com.klp.hub.hub.presentation.dto.response;
+package com.klp.hub.hub.presentation.dto.response.hub;
 
 import java.util.UUID;
 
-public record UpdatedHubResponse(
+public record GetHubDetailResponse(
     UUID hubId,
     String name,
     Long latitude,
