+++ hub/src/main/java/com/klp/hub/hub/presentation/dto/response/hub/GetHubListResponse.java
@@ -1,4 +1,4 @@
-package com.klp.hub.hub.presentation.dto.response;
+package com.klp.hub.hub.presentation.dto.response.hub;
 
 import java.util.List;
 import java.util.UUID;
